fix translation errors
