+++ src/contexts/translations.ts
@@ -1,0 +1,248 @@
+export type Language = 'en' | 'kk';
+
+export type TranslationKey = 
+    | 'login'
+    | 'logout'
+    | 'language'
+    | 'english'
+    | 'kazakh'
+    // Navigation
+    | 'nav.blocks'
+    | 'nav.rooms'
+    | 'nav.tenants'
+    | 'nav.maintenance'
+    | 'nav.reports'
+    // Buildings Page
+    | 'buildings.title'
+    | 'buildings.sortBy'
+    | 'buildings.buildingNumber'
+    | 'buildings.occupancyHighest'
+    | 'buildings.totalRoomsHighest'
+    | 'buildings.loading'
+    | 'buildings.noData'
+    | 'buildings.error'
+    | 'buildings.floors'
+    | 'buildings.totalRooms'
+    | 'buildings.occupied'
+    | 'buildings.occupancy'
+    // Rooms Page
+    | 'rooms.title'
+    | 'rooms.loading'
+    | 'rooms.error.fetch'
+    | 'rooms.error.filters'
+    | 'rooms.filters.building'
+    | 'rooms.filters.status'
+    | 'rooms.filters.bedrooms'
+    | 'rooms.filters.all'
+    | 'rooms.status.available'
+    | 'rooms.status.occupied'
+    | 'rooms.status.maintenance'
+    | 'rooms.table.number'
+    | 'rooms.table.building'
+    | 'rooms.table.type'
+    | 'rooms.table.bedrooms'
+    | 'rooms.table.status'
+    | 'rooms.table.tenant'
+    | 'rooms.table.actions'
+    | 'rooms.table.noData'
+    | 'rooms.table.floor'
+    | 'rooms.table.capacity'
+    | 'rooms.table.area'
+    // Maintenance
+    | 'maintenance.management'
+    | 'maintenance.newRequest'
+    | 'maintenance.cancel'
+    | 'maintenance.room'
+    | 'maintenance.tenant'
+    | 'maintenance.category'
+    | 'maintenance.priority'
+    | 'maintenance.description'
+    | 'maintenance.notes'
+    | 'maintenance.submit'
+    | 'maintenance.submitting'
+    | 'maintenance.filters'
+    | 'maintenance.allStatuses'
+    | 'maintenance.allCategories'
+    | 'maintenance.allPriorities'
+    | 'maintenance.requestDetails'
+    | 'maintenance.requestId'
+    | 'maintenance.status'
+    | 'maintenance.submittedDate'
+    | 'maintenance.assignedTo'
+    | 'maintenance.scheduledDate'
+    | 'maintenance.completedDate'
+    | 'maintenance.updateRequest'
+    | 'maintenance.noChange'
+    | 'maintenance.requestHistory'
+    | 'maintenance.noUpdates'
+    | 'maintenance.updatedBy'
+    | 'maintenance.view'
+    | 'maintenance.notAssigned'
+    | 'maintenance.notScheduled'
+    | 'maintenance.notCompleted'
+    | 'maintenance.unknown';
+
+export interface Translations {
+    [key: string]: {
+        [K in TranslationKey]: string;
+    };
+}
+
+export const translations: Translations = {
+    en: {
+        'login': 'Login',
+        'logout': 'Logout',
+        'language': 'Language',
+        'english': 'English',
+        'kazakh': 'Kazakh',
+        'nav.blocks': 'Blocks',
+        'nav.rooms': 'Rooms',
+        'nav.tenants': 'Tenants',
+        'nav.maintenance': 'Maintenance',
+        'nav.reports': 'Reports',
+        'buildings.title': 'Blocks',
+        'buildings.sortBy': 'Sort by:',
+        'buildings.buildingNumber': 'Building Number',
+        'buildings.occupancyHighest': 'Occupancy (Highest)',
+        'buildings.totalRoomsHighest': 'Total Rooms (Highest)',
+        'buildings.loading': 'Loading buildings...',
+        'buildings.noData': 'No buildings found.',
+        'buildings.error': 'Failed to fetch building data',
+        'buildings.floors': 'Floors',
+        'buildings.totalRooms': 'Total Rooms',
+        'buildings.occupied': 'Occupied',
+        'buildings.occupancy': 'Occupancy',
+        'rooms.title': 'Rooms',
+        'rooms.loading': 'Loading rooms...',
+        'rooms.error.fetch': 'Failed to fetch room data',
+        'rooms.error.filters': 'Could not load filter options.',
+        'rooms.filters.building': 'Building',
+        'rooms.filters.status': 'Status',
+        'rooms.filters.bedrooms': 'Bedrooms',
+        'rooms.filters.all': 'All',
+        'rooms.status.available': 'Available',
+        'rooms.status.occupied': 'Occupied',
+        'rooms.status.maintenance': 'Under Maintenance',
+        'rooms.table.number': 'Room Number',
+        'rooms.table.building': 'Building',
+        'rooms.table.type': 'Type',
+        'rooms.table.bedrooms': 'Bedrooms',
+        'rooms.table.status': 'Status',
+        'rooms.table.tenant': 'Tenant',
+        'rooms.table.actions': 'Actions',
+        'rooms.table.noData': 'No rooms found',
+        'rooms.table.floor': 'Floor',
+        'rooms.table.capacity': 'Capacity',
+        'rooms.table.area': 'Area',
+        'maintenance.management': 'Maintenance Management',
+        'maintenance.newRequest': 'New Maintenance Request',
+        'maintenance.cancel': 'Cancel',
+        'maintenance.room': 'Room',
+        'maintenance.tenant': 'Tenant',
+        'maintenance.category': 'Category',
+        'maintenance.priority': 'Priority',
+        'maintenance.description': 'Description',
+        'maintenance.notes': 'Additional Notes',
+        'maintenance.submit': 'Submit',
+        'maintenance.submitting': 'Submitting...',
+        'maintenance.filters': 'Filters',
+        'maintenance.allStatuses': 'All Statuses',
+        'maintenance.allCategories': 'All Categories',
+        'maintenance.allPriorities': 'All Priorities',
+        'maintenance.requestDetails': 'Maintenance Request Details',
+        'maintenance.requestId': 'Request ID',
+        'maintenance.status': 'Status',
+        'maintenance.submittedDate': 'Submitted Date',
+        'maintenance.assignedTo': 'Assigned To',
+        'maintenance.scheduledDate': 'Scheduled Date',
+        'maintenance.completedDate': 'Completed Date',
+        'maintenance.updateRequest': 'Update Request',
+        'maintenance.noChange': 'No Change',
+        'maintenance.requestHistory': 'Request History',
+        'maintenance.noUpdates': 'No updates found',
+        'maintenance.updatedBy': 'Updated by',
+        'maintenance.view': 'View',
+        'maintenance.notAssigned': 'Not assigned',
+        'maintenance.notScheduled': 'Not scheduled',
+        'maintenance.notCompleted': 'Not completed',
+        'maintenance.unknown': 'Unknown'
+    },
+    kk: {
+        'login': 'Кіру',
+        'logout': 'Шығу',
+        'language': 'Тіл',
+        'english': 'Ағылшын',
+        'kazakh': 'Қазақ',
+        'nav.blocks': 'Блоктар',
+        'nav.rooms': 'Бөлмелер',
+        'nav.tenants': 'Тұрғындар',
+        'nav.maintenance': 'Жөндеу қызметтері',
+        'nav.reports': 'Репорттар',
+        'buildings.title': 'Блоктар',
+        'buildings.sortBy': 'Сұрыптау:',
+        'buildings.buildingNumber': 'Блок нөірі',
+        'buildings.occupancyHighest': 'Толымдылық (Жоғары)',
+        'buildings.totalRoomsHighest': 'Бөлмелер саны (Жоғары)',
+        'buildings.loading': 'Блоктар жүктелуде...',
+        'buildings.noData': 'Блоктар табылмады.',
+        'buildings.error': 'Блок мәліметтерін жүктеу қатесі',
+        'buildings.floors': 'Қабаттар',
+        'buildings.totalRooms': 'Барлық бөлмелер',
+        'buildings.occupied': 'Бос емес',
+        'buildings.occupancy': 'Толымдылық',
+        'rooms.title': 'Бөлмелер',
+        'rooms.loading': 'Бөлмелер жүктелуде...',
+        'rooms.error.fetch': 'Бөлме мәліметтерін жүктеу қатесі',
+        'rooms.error.filters': 'Сүзгі параметрлерін жүктеу мүмкін емес.',
+        'rooms.filters.building': 'Ғимарат',
+        'rooms.filters.status': 'Күйі',
+        'rooms.filters.bedrooms': 'Жатын бөлмелер',
+        'rooms.filters.all': 'Барлық',
+        'rooms.status.available': 'Бос',
+        'rooms.status.occupied': 'Бос емес',
+        'rooms.status.maintenance': 'Жөндеуде',
+        'rooms.table.number': 'Бөлме нөмірі',
+        'rooms.table.building': 'Ғимарат',
+        'rooms.table.type': 'Түрі',
+        'rooms.table.bedrooms': 'Жатын бөлмелер',
+        'rooms.table.status': 'Күйі',
+        'rooms.table.tenant': 'Тұрғын',
+        'rooms.table.actions': 'Әрекеттер',
+        'rooms.table.noData': 'Бөлмелер табылмады',
+        'rooms.table.floor': 'Қабат',
+        'rooms.table.capacity': 'Сыйымдылық',
+        'rooms.table.area': 'Ауданы',
+        'maintenance.management': 'Жөндеу Басқаруы',
+        'maintenance.newRequest': 'Жаңа Жөндеу Сұрауы',
+        'maintenance.cancel': 'Болдырмау',
+        'maintenance.room': 'Бөлме',
+        'maintenance.tenant': 'Тұрғын',
+        'maintenance.category': 'Санат',
+        'maintenance.priority': 'Басымдық',
+        'maintenance.description': 'Сипаттама',
+        'maintenance.notes': 'Қосымша ескертулер',
+        'maintenance.submit': 'Жіберу',
+        'maintenance.submitting': 'Жіберілуде...',
+        'maintenance.filters': 'Сүзгілер',
+        'maintenance.allStatuses': 'Барлық Статустар',
+        'maintenance.allCategories': 'Барлық Санаттар',
+        'maintenance.allPriorities': 'Барлық Басымдықтар',
+        'maintenance.requestDetails': 'Жөндеу Сұрауының Толық Ақпараты',
+        'maintenance.requestId': 'Сұрау ID',
+        'maintenance.status': 'Статус',
+        'maintenance.submittedDate': 'Жіберілген Күні',
+        'maintenance.assignedTo': 'Тапсырылған',
+        'maintenance.scheduledDate': 'Жоспарланған Күні',
+        'maintenance.completedDate': 'Аяқталған Күні',
+        'maintenance.updateRequest': 'Сұрауды Жаңарту',
+        'maintenance.noChange': 'Өзгеріс жоқ',
+        'maintenance.requestHistory': 'Сұрау Тарихы',
+        'maintenance.noUpdates': 'Жаңартулар табылмады',
+        'maintenance.updatedBy': 'Жаңартқан',
+        'maintenance.view': 'Қарау',
+        'maintenance.notAssigned': 'Тапсырылмаған',
+        'maintenance.notScheduled': 'Жоспарланбаған',
+        'maintenance.notCompleted': 'Аяқталмаған',
+        'maintenance.unknown': 'Белгісіз'
+    }
+};
