--- src/pages/Login.module.css
@@ -1,112 +1,0 @@
-.login-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  background-color: #f9fafb;
-  background-image: linear-gradient(135deg, #f5f7fa 0%, #c3cfe2 100%);
-}
-
-.login-card {
-  background: white;
-  border-radius: 8px;
-  box-shadow: 0 10px 25px rgba(0, 0, 0, 0.1);
-  width: 100%;
-  max-width: 400px;
-  padding: 2rem;
-}
-
-.login-header {
-  text-align: center;
-  margin-bottom: 2rem;
-}
-
-.login-header h1 {
-  margin: 0 0 0.5rem 0;
-  color: #2d3748;
-  font-size: 1.75rem;
-}
-
-.login-header p {
-  margin: 0;
-  color: #718096;
-  font-size: 1rem;
-}
-
-.login-form {
-  margin-bottom: 1.5rem;
-}
-
-.login-form .form-group {
-  margin-bottom: 1.25rem;
-}
-
-.login-form input {
-  width: 100%;
-  padding: 0.75rem;
-  border: 1px solid #e2e8f0;
-  border-radius: 0.375rem;
-  font-size: 1rem;
-  transition: all 0.2s;
-}
-
-.login-form input:focus {
-  outline: none;
-  border-color: #4299e1;
-  box-shadow: 0 0 0 3px rgba(66, 153, 225, 0.1);
-}
-
-.login-form button {
-  width: 100%;
-  padding: 0.75rem;
-  background-color: #4299e1;
-  color: white;
-  border: none;
-  border-radius: 0.375rem;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.login-form button:hover {
-  background-color: #3182ce;
-}
-
-.login-form button:disabled {
-  background-color: #a0aec0;
-  cursor: not-allowed;
-}
-
-.login-error {
-  padding: 0.75rem;
-  margin-bottom: 1rem;
-  background-color: #fff5f5;
-  border: 1px solid #fed7d7;
-  border-radius: 0.375rem;
-  color: #c53030;
-  font-size: 0.875rem;
-}
-
-.login-help {
-  margin-top: 2rem;
-  padding-top: 1rem;
-  border-top: 1px solid #e2e8f0;
-  font-size: 0.875rem;
-  color: #718096;
-  text-align: center;
-}
-
-.login-help p {
-  margin: 0.25rem 0;
-}
-
-.login-help a {
-  color: #4299e1;
-  text-decoration: none;
-  font-weight: 500;
-}
-
-.login-help a:hover {
-  text-decoration: underline;
-}
